12.12.2023

+++ DbConnection/Demirciler/MSSQL/Business/DataBaseObjects1.cs
@@ -227,7 +227,7 @@ namespace Projects.DbConnection.Business.MSSQL
         public int? p_id { get; set; }
         public string giris_zaman { get; set; }
         public string cikis_zaman { get; set; }
-        public int? mesai_sure { get; set; }
+        public double? mesai_sure { get; set; }
         public string tarih { get; set; }
         public string c_sure { get; set;}
         public decimal? ucret { get; set; }
